Revamped generator system
It's still a little mucky, but overall I think it's a lot better than it was when I started work on this.

At first I had a very, very bad setup that would've made it hard to add any sort of generator, now generators are
(hopefully) pretty easy to make.

I plan to improve this in the future, probably add a folder containing the main generator files (gen.h, platform layers, etc.)

BUGFIX: The data generator was generating the wrong offsets for pak files, so I updated it to generate the correct offsets
        I also updated the generator system a bit more, giving the generators a working directory and an output directory to dump their files.